외화 송금 99% 완

+++ flobank-api/src/main/java/kr/co/api/flobankapi/dto/FrgnAcctBalanceDTO.java
@@ -11,7 +11,7 @@ public class FrgnAcctBalanceDTO {
 
     String balNo;
     String balCurrency;
-    Integer balBalance;
+    double balBalance;
     String balRegDt;
 
     String balFrgnAcctNo; // 모체 통장 계좌번호
